requirements.txt
discord.py
python-dotenv
aiohttp
beautifulsoup4
opencv-python
numpy
pytesseract
